TCSApiAccessor/requirements.txt
uvicorn>=0.30.1
fastapi>=0.111.0
tinkoff-investments>=0.2.0b100
